+++ appligator/src/appligator/airflow/gen_workflow_dag.py
@@ -3,6 +3,7 @@
 #  https://opensource.org/license/apache-2-0.
 
 from appligator.airflow.ir import workflow_to_ir
+from appligator.airflow.models import ConfigMapMount, PvcMount, ResourceRequirements
 from appligator.airflow.renderer import AirflowRenderer
 from procodile.workflow import WorkflowStepRegistry
 
@@ -11,6 +12,10 @@ def gen_workflow_dag(
     dag_id: str,
     registry: WorkflowStepRegistry,
     image: str,
+    env_from_secrets: list[str] | None = None,
+    resources: ResourceRequirements | None = None,
+    pvc_mounts: list[PvcMount] | None = None,
+    config_map_mounts: list[ConfigMapMount] | None = None,
 ) -> str:
     """Generates a fully-formed Airflow DAG Python file."""
 
@@ -21,7 +26,15 @@ def gen_workflow_dag(
         raise ValueError("Image name is required to generate dag.")
 
     # operator-agnostic intermediate representation
-    ir = workflow_to_ir(registry, dag_id, image_name=image)
+    ir = workflow_to_ir(
+        registry,
+        dag_id,
+        image_name=image,
+        env_from_secrets=env_from_secrets,
+        resources=resources,
+        pvc_mounts=pvc_mounts,
+        config_map_mounts=config_map_mounts,
+    )
 
     dag_code = AirflowRenderer().render(ir)